requirements.txt
pandas>=2.2
numpy>=1.26
pyarrow
google-cloud-bigquery
lightgbm
optuna
pytorch-lightning
pytorch-forecasting
mlflow
ta
shap
fastapi
uvicorn[standard]
evidently
scikit-learn
PyJWT
python-multipart
passlib[bcrypt]
python-jose[cryptography]
